Add context switch code for MITTS registers
MITTS has two configuration registers that must be swapped on a
context switch. These registers are written by a hypervisor
hcall. Since the time to context switch is massively longer than
our bin refresh rate, there is no reason to save the register
bin values -- although we can implement this later.

The desired bin configuration and settings are stored in the
thread_info structure as a two-entry long int array.